move member function decl/def stuff to class definition file
