+++ src/js/core/constructs/declarations/declarations.ts
@@ -189,110 +189,7 @@ export function createLocalSimpleDeclarationFromAST(ast: NonMemberSimpleDeclarat
     });
 }
 
-const MemberDeclarationConstructsMap = {
-    "simple_member_declaration": (ast: MemberSimpleDeclarationASTNode, context: MemberSpecificationContext) => createMemberSimpleDeclarationFromAST(ast, context),
-    "function_definition": (ast: FunctionDefinitionASTNode, context: MemberSpecificationContext) => createMemberFunctionDeclarationFromDefinitionAST(ast, context)
-    // Note: function_definition includes ctor and dtor definitions
-};
 
-export function createMemberDeclarationFromAST<ASTType extends MemberDeclarationASTNode>(ast: ASTType, context: MemberSpecificationContext) : ReturnType<(typeof MemberDeclarationConstructsMap)[ASTType["construct_type"]]>{
-    return <any>MemberDeclarationConstructsMap[ast.construct_type](<any>ast, context);
-}
-
-export function createMemberSimpleDeclarationFromAST(ast: MemberSimpleDeclarationASTNode, context: MemberSpecificationContext) {
-    // assert(isMemberSpecificationContext(context), "A Member declaration must be created in a member specification context.");
-
-    // Need to create TypeSpecifier first to get the base type for the declarators
-    let typeSpec = TypeSpecifier.createFromAST(ast.specs.typeSpecs, context);
-    let baseType = typeSpec.baseType;
-    let storageSpec = StorageSpecifier.createFromAST(ast.specs.storageSpecs, context);
-
-    // A constructor may have been parsed incorrectly due to an ambiguity in the grammar.
-    // For example, A(); might have been parsed as a function returning an A with a declarator
-    // that is missing its name. In that case, A would be the type specifier.
-    // So, we check the first declarator. If it has no name, and the type specifier
-    // identified the contextual class type, we know this mistake has occurred and we fix it.
-    if (baseType?.sameType(context.containingClass.type)) {
-        let testDeclarator = Declarator.createFromAST(ast.declarators[0], context, baseType);
-        if (!testDeclarator.name) {
-            typeSpec = TypeSpecifier.createFromAST(ast.specs.typeSpecs.filter(spec => spec !== context.containingClass.name), context);
-        }
-    }
-
-
-    // Create an array of the individual declarations (multiple on the same line
-    // will be parsed as a single AST node and need to be broken up)
-    return ast.declarators.map((declAST) => {
-
-        // Create declarator and determine declared type
-        let declarator = Declarator.createFromAST(declAST, context, baseType);
-        let declaredType = declarator.type;
-
-        // Create the declaration itself. Which kind depends on the declared type
-        let declaration: MemberSimpleDeclaration;
-        if (!declaredType) {
-            declaration = new UnknownTypeDeclaration(context, ast, typeSpec, storageSpec, declarator, ast.specs);
-        }
-        else if (ast.specs.friend) {
-            declaration = new FriendDeclaration(context, ast, typeSpec, storageSpec, declarator, ast.specs);
-        }
-        else if (ast.specs.typedef) {
-            declaration = new TypedefDeclaration(context, ast, typeSpec, storageSpec, declarator, ast.specs);
-        }
-        else if (declaredType.isVoidType()) {
-            declaration = new VoidDeclaration(context, ast, typeSpec, storageSpec, declarator, ast.specs);
-        }
-        else if (declaredType.isFunctionType()) {
-            declaration = new FunctionDeclaration(context, ast, typeSpec, storageSpec, declarator, ast.specs, declaredType);
-        }
-        else if (declaredType.isArrayOfUnknownBoundType()) {
-            // TODO: it may be possible to determine the bound from the initializer
-            declaration = new UnknownBoundArrayDeclaration(context, ast, typeSpec, storageSpec, declarator, ast.specs, declaredType);
-        }
-        else if (declaredType.isCompleteObjectType() || declaredType.isReferenceType()) {
-            declaration = new MemberVariableDeclaration(context, ast, typeSpec, storageSpec, declarator, ast.specs, declaredType);
-            if (declAST.initializer) {
-                // member variables don't get anything set for a default initializer,
-                // so this if keeps us from doing anything unless there's an explicit
-                // initialization in the AST
-                setInitializerFromAST(declaration, declAST.initializer, context);
-            }
-        }
-        else {
-            declaration = new IncompleteTypeMemberVariableDeclaration(context, ast, typeSpec, storageSpec, declarator, ast.specs, declaredType);
-        }
-
-        return declaration;
-    });
-}
-
-export function createMemberFunctionDeclarationFromDefinitionAST(ast: FunctionDefinitionASTNode, context: TranslationUnitContext) {
-
-    // Need to create TypeSpecifier first to get the base type for the declarators
-    const typeSpec = TypeSpecifier.createFromAST(ast.specs.typeSpecs, context);
-    const baseType = typeSpec.baseType;
-    const storageSpec = StorageSpecifier.createFromAST(ast.specs.storageSpecs, context);
-
-    const declarator = Declarator.createFromAST(ast.declarator, context, baseType);
-    const declaredType = declarator.type;
-
-    if (!declarator.name) {
-        return new InvalidConstruct(context, ast, CPPError.declaration.missing_name);
-    }
-    
-    if (!declaredType?.isFunctionType()) {
-        return new InvalidConstruct(context, ast, CPPError.declaration.func.definition_non_function_type);
-    }
-    
-    const declAST: SimpleDeclarationASTNode = {
-        construct_type: "simple_declaration",
-        declarators: [ast.declarator],
-        specs: ast.specs,
-        source: ast.declarator.source
-    };
-
-    return new FunctionDeclaration(context, declAST, typeSpec, storageSpec, declarator, ast.specs, declaredType);
-}
